Clear the stats cache before getting the file size
We are hitting this bug http://bugs.php.net/bug.php?id=28790 since the
stats are cached when the file has been changed and doing a reupload it
will get the same file size as before which would generate a 500 when
uploading to Rackspace Cloud Files.

I am not entirely sure about the performances issue of clearing all the
stats cache but since we are only playing safe I am adding the
clearstatecache().